requirements.txt
# ═══════════════════════════════════════════
# LoRA Data Toolkit — Core Dependencies
# ═══════════════════════════════════════════
# These are needed to run the app.
# For LoRA training deps, use the in-app
# Setup page (or see README.md).
# ═══════════════════════════════════════════

# GUI
customtkinter>=5.2.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
trafilatura>=1.6.0
lxml>=4.9.0

# YouTube
youtube-transcript-api>=0.6.0

# OCR + Images
pytesseract>=0.3.10
Pillow>=10.0.0

# PDF
PyMuPDF>=1.23.0

# System tray + Hotkeys + Drag-and-drop
pystray>=0.19.0
keyboard>=0.13.0
windnd>=0.1.0

# Config / Merge
pyyaml>=6.0
